Add PostgreSQL-compatible hybrid bitmapset with dual-mode storage
Combine chunk_codec.h and bitmapset_hybrid.[ch] into PostgreSQL
drop-in replacement files (test/bitmapset.[ch]) with a postgres_compat.h
shim for standalone testing. The hybrid bitmapset uses dense word-array
storage for sets up to 262K bits (BMS_DENSE_MAX_NWORDS=4096) and
automatically transitions to RLE+sparse chunk encoding for larger sets.

Key changes:
- test/bitmapset.[ch]: PG-compatible API with int64_t member indices,
  palloc/pfree memory management, elog error reporting, and tab-indented
  PG coding style. All 30 public API functions support both storage modes.
- test/postgres_compat.h: Shim mapping palloc->malloc, elog->fprintf+abort,
  Assert->assert for building outside PostgreSQL.
- include/chunk_codec.h: Sparsemap chunk codec (RLE + sparse encoding)
  extracted as a standalone header.
- test/bitmapset_hybrid.[ch]: Original hybrid implementation (source for
  the PG conversion).
- test/bench_hybrid_modes.c: Benchmark for offset sweep (dense->chunked
  transition) and cutoff sweep (optimal BMS_DENSE_MAX_NWORDS).
- test/bench_hybrid_wrapper.c: Unity-build wrapper with symbol renaming
  for linking hybrid bitmapset alongside bitmapset_standalone in bench.

+++ include/sparsemap.h
@@ -439,7 +439,7 @@ uint64_t sparsemap_span(sparsemap_t *map, uint64_t start, size_t len, bool value
  *   sparsemap_scan(map, print_bits, 0, NULL);
  * @endcode
  */
-void sparsemap_scan(const sparsemap_t *map, void (*scanner)(uint32_t vec[], size_t n, void *aux), size_t skip, void *aux);
+void sparsemap_scan(const sparsemap_t *map, void (*scanner)(uint64_t vec[], size_t n, void *aux), size_t skip, void *aux);
 
 /* -------------------------------------------------------------------
  * Bulk operations
