+++ test/test_rle_standalone.c
@@ -3,6 +3,7 @@
  * Compile: cc -Wall -Wextra -g -I. -o test_rle test_rle_standalone.c sparsemap.c
  */
 
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <assert.h>
@@ -10,10 +11,10 @@
 
 /* Test counter for scan */
 static size_t scan_count = 0;
-static uint32_t scan_last_idx = 0;
+static uint64_t scan_last_idx = 0;
 
 void
-scan_counter(uint32_t v[], size_t n, void *aux)
+scan_counter(uint64_t v[], size_t n, void *aux)
 {
   (void)aux;
   for (size_t i = 0; i < n; i++) {
